feat#86a79vyxn/Filtrar usuários na query

+++ src/main/java/com/reportai/reportaiserver/dto/UsuariosPaginadoDTO.java
@@ -15,6 +15,6 @@ public class UsuariosPaginadoDTO {
    private int pagina;
    private int limite;
    private int totalPaginas;
-   private Long totalUsuarios;
+   private int totalUsuarios;
    private List<UsuarioDTO> usuarios;
 }
